+++ lib/src/handler.dart
@@ -1,14 +1,28 @@
 part of pwa_worker;
 
 /// Handles [request] and returns a Future which completes with a [Response].
+typedef Future<Response> RequestHandler(Request request);
+
+/// Handles [request] and returns a Future which completes with a [Response].
+@Deprecated('Use RequestHandler instead. Handler will be removed in 0.1')
 typedef Future<Response> Handler(Request request);
 
-/// The default fetch [Handler]: fetching over the network.
-final Handler defaultFetchHandler = fetch;
+/// Network [RequestHandler] with the default disk caching.
+final RequestHandler defaultRequestHandler = fetch;
 
-/// Network handler that forces a network fetch, and skips disk cache.
-final Handler noCacheNetworkFetch =
+/// Network [RequestHandler] with the default disk caching.
+@Deprecated(
+    'Use defaultRequestHandler instead. defaultFetchHandler will be removed in 0.1')
+final RequestHandler defaultFetchHandler = defaultRequestHandler;
+
+/// Network [RequestHandler] that skips the default disk cache.
+final RequestHandler noCacheNetworkRequestHandler =
     (Request request) => fetch(request, new RequestInit(cache: 'no-store'));
+
+/// Network [RequestHandler] that skips disk cache.
+@Deprecated(
+    'Use noCacheNetworkRequestHandler instead. noCacheNetworkFetch will be removed in 0.1')
+final RequestHandler noCacheNetworkFetch = noCacheNetworkRequestHandler;
 
 /// Whether the [response] is valid (e.g. not an error, not a missing item).
 bool isValidResponse(Response response) {
@@ -17,11 +31,12 @@ bool isValidResponse(Response response) {
   return true;
 }
 
-/// Return a composite [Handler] that joins the [handlers] in serial processing,
+/// Return a composite [RequestHandler] that joins the [handlers] in serial processing,
 /// completing with the first valid response. If none of the [handlers]
 /// provide a valid response, it will complete with an error.
-Handler joinHandlers(List<Handler> handlers) => (Request request) async {
-      for (Handler handler in handlers) {
+RequestHandler joinHandlers(List<RequestHandler> handlers) =>
+    (Request request) async {
+      for (RequestHandler handler in handlers) {
         try {
           Response response = await handler(request.clone());
           if (isValidResponse(response)) return response;
@@ -30,10 +45,11 @@ Handler joinHandlers(List<Handler> handlers) => (Request request) async {
       return new Response.error();
     };
 
-/// Returns a composite [Handler] that runs the [handlers] in parallel and
+/// Returns a composite [RequestHandler] that runs the [handlers] in parallel and
 /// completes with the first valid response. If none of the [handlers]
 /// provide a valid response, it will complete with an error.
-Handler raceHandlers(List<Handler> handlers) => (Request request) {
+RequestHandler raceHandlers(List<RequestHandler> handlers) =>
+    (Request request) {
       Completer<Response> completer = new Completer();
       int remaining = handlers.length;
       final complete = (Response response) {
@@ -47,7 +63,7 @@ Handler raceHandlers(List<Handler> handlers) => (Request request) {
           completer.complete(new Response.error());
         }
       };
-      for (Handler handler in handlers) {
+      for (RequestHandler handler in handlers) {
         handler(request.clone()).then((Response response) {
           complete(response);
         }, onError: (e) {